Correção do model de produto e termino dos controllers

+++ Controllers/GerenteController.cs
@@ -17,49 +17,27 @@ namespace MercadoApi.Controllers
         public GerenteController(Context context)
         {
             _context = context;
-            
-            if (_context.Gerentes.Count() == 0)
-            {
-                // Create a new TodoItem if collection is empty,
-                // which means you can't delete all TodoItems.
-                var gerente = new Gerente {Funcionario = new Funcionario{Nome = "Gerente", Cpf = "1",Sexo = "M", Turno = "Manhã,Tarde,Noite", Salario =  5000.00}};
-                _context.Gerentes.Add(gerente);
-                _context.SaveChanges();
-            }
         }
 
-    //SSL
     // GET: 
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<Object>>> GetGerentes()
+    public async Task<ActionResult<IEnumerable<Gerente>>> GetGerentes()
     {
-        var gerenteList = await _context.Gerentes.ToListAsync();
-        Object[] newGerenteList = new Object[gerenteList.Count];
-        int i = 0;
-        foreach(var gerente in gerenteList){
-            var funcionario = await _context.Funcionarios.FindAsync(gerente.FuncionarioId);
-            var json = new{
-                id = gerente.Id,
-                funcionario =  funcionario
-            };
-            System.Diagnostics.Debug.WriteLine(json);
-            newGerenteList[i] = json;
-            i++;
-        }
+        var gerenteList = await _context.Gerentes.Include(gerente => gerente.Funcionario).ToListAsync();
        
-        return  newGerenteList.ToList();
+        return  gerenteList;
     }
 
     // GET with id: 
     [HttpGet("{id}")]
-    public async Task<ActionResult<Object>> GetGerentes(int id)
+    public async Task<ActionResult<Object>> GetGerente(int id)
     {
         var gerente = await _context.Gerentes.FindAsync(id);
-        var funcionario = await _context.Funcionarios.FindAsync(gerente.FuncionarioId);
         if (gerente == null)
         {
             return NotFound();
         }
+        var funcionario = await _context.Funcionarios.FindAsync(gerente.FuncionarioId);
         var gerenteFormatado = new {
             id = gerente.Id,
             funcionario = funcionario
@@ -79,30 +57,35 @@ namespace MercadoApi.Controllers
 
     // PUT: 
     [HttpPut("{id}")]
-    public async Task<IActionResult> PutGerentes(long id, Gerente item)
+    public async Task<IActionResult> PutGerente(int id, Gerente item)
     {
         if (id != item.Id)
         {
             return BadRequest();
         }
-
+        _context.Funcionarios.Update(item.Funcionario);
         _context.Entry(item).State = EntityState.Modified;
         await _context.SaveChangesAsync();
 
         return NoContent();
     }
 
+
     // DELETE with id:
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteTodoItem(int id)
+    public async Task<IActionResult> DeleteGerente(int id)
     {
         var gerente = await _context.Gerentes.FindAsync(id);
-
         if (gerente == null)
         {
             return NotFound();
         }
-
+        var funcionario = await _context.Funcionarios.FindAsync(gerente.FuncionarioId);
+        if (funcionario == null)
+        {
+            return NotFound();
+        }
+        _context.Funcionarios.Remove(funcionario);
         _context.Gerentes.Remove(gerente);
         await _context.SaveChangesAsync();
